+++ libsmartcols/src/table.c
@@ -332,7 +332,7 @@ int scols_table_next_column(struct libscols_table *tb,
  */
 int scols_table_get_ncols(struct libscols_table *tb)
 {
-	return tb ? tb->ncols : -EINVAL;
+	return tb ? (int)tb->ncols : -EINVAL;
 }
 
 /**
@@ -343,7 +343,7 @@ int scols_table_get_ncols(struct libscols_table *tb)
  */
 int scols_table_get_nlines(struct libscols_table *tb)
 {
-	return tb ? tb->nlines : -EINVAL;
+	return tb ? (int)tb->nlines : -EINVAL;
 }
 
 /**
